[Lightning WIP] Sublass from Module instead of contain
