+++ configs/first_working_lightning_config/ml/model.yml
@@ -1,16 +1,13 @@
-model:
-  class_path: gnn_tracking.models.graph_construction.GraphConstructionFCNN
-  init_args:
-    in_dim: 14
-    out_dim: 8
-    hidden_dim: 512
-    depth: 5
-loss_fct:
-  class_path: gnn_tracking.metrics.losses.GraphConstructionHingeEmbeddingLoss
-  init_args:
-    p_attr: 2.
-optimizer:
-  class_path: torch.optim.Adam
-  init_args:
-    lr: 0.0001
-    weight_decay: 0.0001
+class_path: gnn_tracking.models.graph_construction.GraphConstructionFCNN
+init_args:
+  in_dim: 14
+  out_dim: 8
+  hidden_dim: 512
+  depth: 5
+  loss_cfg:
+      p_attr: 2.
+  optimizer:
+    class_path: torch.optim.Adam
+    init_args:
+      lr: 0.0001
+      weight_decay: 0.0001
